feat: implemented second task

+++ Inventory/Domain/Order/OrderItem/ValueObjects/OrderItemPrice.cs
@@ -6,16 +6,16 @@ namespace Inventory.Domain.Order.OrderItem.ValueObjects;
 
 public class OrderItemPrice : ValueObject
 {
-    public int Value { get; private set; }
+    public double Value { get; private set; }
     public Currency Currency { get; private set; }
     
-    private OrderItemPrice(int value, Currency currency)
+    private OrderItemPrice(double value, Currency currency)
     {
         Value = value;
         Currency = currency;
     }
 
-    public static ErrorOr<OrderItemPrice> Create(int value, Currency currency)
+    public static ErrorOr<OrderItemPrice> Create(double value, Currency currency)
     {
         if (value < 0)
             return Error.Validation(OrderItemErrors.OrderItemPriceCanNotBeNegative);
